requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==1.3.0
Cython==0.29.30
docker-pycreds==0.4.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.3
gym==0.17.2
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
importlib-resources==5.8.0
joblib==1.1.0
lockfile==0.12.2
matplotlib==3.5.3
#mujoco-py==1.50.1.68
numpy==1.23.1
omegaconf==2.2.2
packaging==21.3
pathtools==0.1.2
Pillow==9.2.0
promise==2.3
protobuf==3.20.1
psutil==5.9.1
pycparser==2.21
pyglet==1.5.0
pyparsing==3.0.9
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
sentry-sdk==1.6.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
torch==1.12.0
typing-extensions==4.3.0
urllib3==1.26.10
wandb==0.12.21
zipp==3.8.0
